+++ src/main/kotlin/net/pwall/json/schema/codegen/CodeGenerator.kt
@@ -774,6 +774,21 @@ class CodeGenerator(
                     property.addValidation(Validation.Type.IPV6)
                     return true
                 }
+                FormatValidator.DurationFormatChecker.name -> {
+                    target.systemClasses.addOnce(SystemClass.VALIDATION)
+                    property.addValidation(Validation.Type.DURATION)
+                    return true
+                }
+                FormatValidator.JSONPointerFormatChecker.name -> {
+                    target.systemClasses.addOnce(SystemClass.VALIDATION)
+                    property.addValidation(Validation.Type.JSON_POINTER)
+                    return true
+                }
+                FormatValidator.RelativeJSONPointerFormatChecker.name -> {
+                    target.systemClasses.addOnce(SystemClass.VALIDATION)
+                    property.addValidation(Validation.Type.RELATIVE_JSON_POINTER)
+                    return true
+                }
                 FormatValidator.DateTimeFormatChecker.name -> {
                     target.systemClasses.addOnce(SystemClass.DATE_TIME)
                     property.systemClass = SystemClass.DATE_TIME
@@ -785,10 +800,6 @@ class CodeGenerator(
                 FormatValidator.TimeFormatChecker.name -> {
                     target.systemClasses.addOnce(SystemClass.TIME)
                     property.systemClass = SystemClass.TIME
-                }
-                FormatValidator.DurationFormatChecker.name -> {
-                    target.systemClasses.addOnce(SystemClass.DURATION)
-                    property.systemClass = SystemClass.DURATION
                 }
                 FormatValidator.UUIDFormatChecker.name -> {
                     target.systemClasses.addOnce(SystemClass.UUID)